Add GetPokemon implementation and test suite
Add the new app folder that is the implementation for the app based off
of the domain module. This way is a little simpler and more obvious than
having the domain stand alone with just entities and entity gateways.:

This commit add an implementation of the GetPokemonInteractor which is a
renaming of the original interface for the GetPokemon interactor class.
We also add test suite with three tests for this class.