Fix clear button and empty search functionality (#9)
This commit fixes the issues seen where the clear button shows up when
we don't want it to. We also fix some problems with the search
functionality. Now, if the user leaves an empty search, we load all
Pokemon. If the search is changed from a valid search to a search with 1
character, we do nothing.